Refactor PaginatedQueue into a separate library file.
